pyproject.toml
[tool.poetry]
name = "repl_nix_galileo"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
fredapi = "^0.5.0"
pandas = "^1.5.1"
python-dotenv = "^0.21.0"
Flask = "^2.2.2"
PyYAML = "^6.0"
plotly = "^5.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
